PRJ2, List Done! Now It's time to hash!

+++ src/modules/memory.h
@@ -7,8 +7,10 @@
 	#define FFFF0_8 1048560
 	#define TEN_LINES hex2int("9F")  // needs "core/hex.h".
 
+	#include <stdbool.h>
+
 	size_t memory_dump(void *memory, size_t start, size_t end);
-	bool memory_edit(void *memory, size_t address, int value);
-	bool memory_fill(void *memory, size_t start, size_t end, int value);
+	bool memory_edit(void *memory, size_t address, size_t value);
+	bool memory_fill(void *memory, size_t start, size_t end, size_t value);
 	void memory_reset(void *memory);
 #endif
